+++ src/prop.lalrpop
@@ -1,30 +1,32 @@
 // Bah humbug the auto-generated sources have comments above this comment.
 // #![allow(clippy::all)]
 use crate::lex;
+use crate::token_wrap;
 use crate::ast::{Prop, Expr, Binding, Typ};
 use std::rc::Rc;
+
 grammar<'a>;
 
 extern {
   type Location = usize;
   type Error = lex::LexicalError;
 
-  enum lex::Token<'a> {
-	"⊥" => lex::Token::Bot,
-	"." => lex::Token::Dot,
-	"≔" => lex::Token::Def,
-	"→" => lex::Token::Arrow,
-	"↔" => lex::Token::Iff,
-	"¬" => lex::Token::Neg,
-	"ⲗ" => lex::Token::Abs,
-	"∧" => lex::Token::Conj,
-	"∨" => lex::Token::Disj,
-	"⊤" => lex::Token::Top,
-	"(" => lex::Token::LParen,
-	")" => lex::Token::RParen,
-	":" => lex::Token::Colon,
-	";" => lex::Token::Semi,
-	name => lex::Token::Name(<&'a str>),
+  enum token_wrap::Token<'a> {
+	"⊥" => token_wrap::Token::Bot,
+	"." => token_wrap::Token::Dot,
+	"≔" => token_wrap::Token::Def,
+	"→" => token_wrap::Token::Arrow,
+	"↔" => token_wrap::Token::Iff,
+	"¬" => token_wrap::Token::Neg,
+	"ⲗ" => token_wrap::Token::Abs,
+	"∧" => token_wrap::Token::Conj,
+	"∨" => token_wrap::Token::Disj,
+	"⊤" => token_wrap::Token::Top,
+	"(" => token_wrap::Token::LParen,
+	")" => token_wrap::Token::RParen,
+	":" => token_wrap::Token::Colon,
+	";" => token_wrap::Token::Semi,
+	name => token_wrap::Token::Name(<&'a str>),
   }
 }
